--- a/internal/store.go
+++ b/internal/store.go
@@ -9,13 +9,13 @@ import (
 	"math/rand"
 	"runtime"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/gammazero/deque"
 	"github.com/zeebo/xxh3"
 
 	"github.com/Yiling-J/theine-go/internal/bf"
+	"github.com/Yiling-J/theine-go/internal/xruntime"
 )
 
 const (
@@ -33,8 +33,15 @@ const (
 )
 
 var (
-	VersionMismatch = errors.New("version mismatch")
+	VersionMismatch      = errors.New("version mismatch")
+	maxStripedBufferSize int
 )
+
+func init() {
+	parallelism := xruntime.Parallelism()
+	roundedParallelism := int(RoundUpPowerOf2(parallelism))
+	maxStripedBufferSize = 4 * roundedParallelism
+}
 
 type Shard[K comparable, V any] struct {
 	hashmap   map[K]*Entry[K, V]
@@ -105,9 +112,9 @@ type Store[K comparable, V any] struct {
 	kvBuilder         func(entry *Entry[K, V]) dequeKV[K, V]
 	policy            *TinyLfu[K, V]
 	timerwheel        *TimerWheel[K, V]
-	readbuf           *Queue[ReadBufItem[K, V]]
+	stripedBuffer     []*Buffer[K, V]
+	mask              uint32
 	cost              func(V) int64
-	readCounter       *atomic.Uint32
 	shards            []*Shard[K, V]
 	cap               uint
 	shardCount        uint
@@ -158,16 +165,21 @@ func NewStore[K comparable, V any](
 		costfn = cost
 	}
 
+	stripedBuffer := make([]*Buffer[K, V], 0, maxStripedBufferSize)
+	for i := 0; i < maxStripedBufferSize; i++ {
+		stripedBuffer = append(stripedBuffer, NewBuffer[K, V]())
+	}
+
 	s := &Store[K, V]{
-		cap:         uint(maxsize),
-		hasher:      hasher,
-		policy:      NewTinyLfu[K, V](uint(policySize), hasher),
-		readCounter: &atomic.Uint32{},
-		readbuf:     NewQueue[ReadBufItem[K, V]](),
-		writebuf:    make(chan WriteBufItem[K, V], writeBufSize),
-		entryPool:   sync.Pool{New: func() any { return &Entry[K, V]{} }},
-		shardCount:  uint(shardCount),
-		doorkeeper:  doorkeeper,
+		cap:           uint(maxsize),
+		hasher:        hasher,
+		policy:        NewTinyLfu[K, V](uint(policySize), hasher),
+		stripedBuffer: stripedBuffer,
+		mask:          uint32(maxStripedBufferSize - 1),
+		writebuf:      make(chan WriteBufItem[K, V], writeBufSize),
+		entryPool:     sync.Pool{New: func() any { return &Entry[K, V]{} }},
+		shardCount:    uint(shardCount),
+		doorkeeper:    doorkeeper,
 		kvBuilder: func(entry *Entry[K, V]) dequeKV[K, V] {
 			return dequeKV[K, V]{
 				k: entry.key,
@@ -205,7 +217,6 @@ func NewStore[K comparable, V any](
 }
 
 func (s *Store[K, V]) getFromShard(key K, hash uint64, shard *Shard[K, V]) (V, bool) {
-	new := s.readCounter.Add(1)
 	shard.mu.RLock()
 	entry, ok := shard.get(key)
 	var value V
@@ -219,22 +230,17 @@ func (s *Store[K, V]) getFromShard(key K, hash uint64, shard *Shard[K, V]) (V, b
 		}
 	}
 	shard.mu.RUnlock()
-	switch {
-	case new < MAX_READ_BUFF_SIZE:
-		var send ReadBufItem[K, V]
-		send.hash = hash
-		if ok {
-			send.entry = entry
-		}
-		s.readbuf.Push(send)
-	case new == MAX_READ_BUFF_SIZE:
-		var send ReadBufItem[K, V]
-		send.hash = hash
-		if ok {
-			send.entry = entry
-		}
-		s.readbuf.Push(send)
-		s.drainRead()
+
+	idx := s.getReadBufferIdx()
+	var send ReadBufItem[K, V]
+	send.hash = hash
+	if ok {
+		send.entry = entry
+	}
+	pb := s.stripedBuffer[idx].Add(send)
+	if pb != nil {
+		s.drainRead(pb.Returned)
+		s.stripedBuffer[idx].Free()
 	}
 	return value, ok
 }
@@ -528,18 +534,13 @@ func (s *Store[K, V]) removeEntry(entry *Entry[K, V], reason RemoveReason) {
 	}
 }
 
-func (s *Store[K, V]) drainRead() {
+func (s *Store[K, V]) drainRead(buffer []ReadBufItem[K, V]) {
 	s.policy.total.Add(MAX_READ_BUFF_SIZE)
 	s.mlock.Lock()
-	for {
-		v, ok := s.readbuf.Pop()
-		if !ok {
-			break
-		}
-		s.policy.Access(v)
+	for _, e := range buffer {
+		s.policy.Access(e)
 	}
 	s.mlock.Unlock()
-	s.readCounter.Store(0)
 }
 
 func (s *Store[K, V]) maintance() {
@@ -650,6 +651,10 @@ func (s *Store[K, V]) Close() {
 	s.cancel()
 	s.mlock.Unlock()
 	close(s.writebuf)
+}
+
+func (s *Store[K, V]) getReadBufferIdx() int {
+	return int(xruntime.Fastrand() & s.mask)
 }
 
 type StoreMeta struct {