--- a/internal/utils.go
+++ b/internal/utils.go
@@ -42,3 +42,18 @@ func (h *Hasher[K]) hash(key K) uint64 {
 	}
 	return xxh3.HashString(strKey)
 }
+
+// RoundUpPowerOf2 is based on https://graphics.stanford.edu/~seander/bithacks.html#RoundUpPowerOf2.
+func RoundUpPowerOf2(v uint32) uint32 {
+	if v == 0 {
+		return 1
+	}
+	v--
+	v |= v >> 1
+	v |= v >> 2
+	v |= v >> 4
+	v |= v >> 8
+	v |= v >> 16
+	v++
+	return v
+}