- use striped lossy ring buffer as read buffer(https://github.com/maypok86/otter/blob/main/internal/lossy/buffer.go)
- improve clock performance(`time.Unix(0, ts).UTC()` -> `time.Unix(0, ts)`, `UTC()` will strip monotonic clock, and make `Since` slow)
- use RBMutex(https://github.com/puzpuzpuz/xsync/blob/main/rbmutex.go) to improve read performance.
- use striped counter(https://github.com/puzpuzpuz/xsync/blob/main/counter.go) to improve read performance.

before:
```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkGetParallel/theine-12   	        39227174	        31.04 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetParallel/ristretto-12         	47108266	        24.54 ns/op	      17 B/op	       1 allocs/op

BenchmarkGetSingleParallel/theine-12      	28667926	        40.72 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetSingleParallel/ristretto-12   	23612684	        47.27 ns/op	      16 B/op	       1 allocs/op
```

after:
```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkGetParallel/theine-12            52328856	        22.56 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetParallel/ristretto-12         50361355	        25.36 ns/op	      17 B/op	       1 allocs/op
BenchmarkGetParallel/otter-12             181070922	         7.306 ns/op	       0 B/op	       0 allocs/op

BenchmarkGetSingleParallel/theine-12      	53619693	        21.98 ns/op	       0 B/op	       0 allocs/op
BenchmarkGetSingleParallel/ristretto-12   	21261020	        52.45 ns/op	      16 B/op	       1 allocs/op
BenchmarkGetSingleParallel/otter-12       	207251763	         5.774 ns/op	       0 B/op	       0 allocs/op
```