@maypok86 2024-08-09
I used benchmarks based on go-cache-benchmark-plus with a few edits.
1. Theine version borrowed from the perf branch. 
2. Added the `IgnoreInternalCost = true` flag for ristretto.
3. Otter version is the latest release.

I think the current version of go-cache-benchmark has too many bugs, so the benchmarks are fixed too.

```go
func BenchmarkGetParallel(b *testing.B) {
	keys := []string{}
	for i := 0; i < 100000; i++ {
		keys = append(keys, fmt.Sprintf("%d", i))
	}
	for _, client := range benchClients {
		client.Init(100000)
		// The filling has been added
		for _, key := range keys {
			client.Set(key, key)
		}
		b.ResetTimer()
		b.Run(client.Name(), func(b *testing.B) {
			b.RunParallel(func(p *testing.PB) {
			        // It seems better not to read one key out of all the goroutines,
			        // but the results of this almost do not change in any way.
				counter := rand.Int() % 100000
				for p.Next() {
					client.Get(keys[counter%100000])
					counter++
				}
			})
		})
		client.Close()
	}
}
```
I got the following on M1 Max.
```
goos: darwin
goarch: arm64
pkg: github.com/Yiling-J/go-cache-benchmark-plus
BenchmarkGetParallel
BenchmarkGetParallel/theine
BenchmarkGetParallel/theine-10         	16778953	        71.68 ns/op
BenchmarkGetParallel/ristretto
BenchmarkGetParallel/ristretto-10      	46205827	        28.45 ns/op
BenchmarkGetParallel/otter
BenchmarkGetParallel/otter-10          	194468833	         6.116 ns/op
PASS
```
Everything is a bit different on Set benchmarks.
```go
func BenchmarkSetParallel(b *testing.B) {
	keys := []string{}
	for i := 0; i < 1000000; i++ {
		keys = append(keys, fmt.Sprintf("%d", i))
	}
	for _, client := range benchClients {
		client.Init(100000)
		b.ResetTimer()
		b.Run(client.Name(), func(b *testing.B) {
			b.RunParallel(func(p *testing.PB) {
				// 1 insert and inf updates seem to be too cheating for theine.
				// Since it has a fastpath, which will only help for this case.
				// So let's add a random selection.
				counter := rand.Int() % 1000000
				for p.Next() {
					client.Set(keys[counter%1000000], "bar")
					counter++
				}
			})
		})
		client.Close()
	}
}
```
And we get the following.
```
goos: darwin
goarch: arm64
pkg: github.com/Yiling-J/go-cache-benchmark-plus
BenchmarkSetParallel
BenchmarkSetParallel/theine
BenchmarkSetParallel/theine-10         	 2895408	       346.8 ns/op
BenchmarkSetParallel/ristretto
BenchmarkSetParallel/ristretto-10      	16717233	       123.5 ns/op
BenchmarkSetParallel/otter
BenchmarkSetParallel/otter-10          	 2526064	       486.4 ns/op
PASS
```