@Yiling-J 2024-08-10
@maypok86 I actually prefill the cache in the Get method, but the code hasn't been pushed yet. I've now updated the repo. The Set benchmark also involves some randomness. I found that I use the total key count as cache size in set, so I've switched to using 1/10th of the total key count. This should increases evictions and show the cost of that. New results:
```
BenchmarkGetParallel/theine-10   	45067902	        27.19 ns/op
BenchmarkGetParallel/ristretto-10         	46896685	        27.42 ns/op
BenchmarkGetParallel/otter-10             	142256046	         8.454 ns/op
BenchmarkGetSingleParallel/theine-10      	48784158	        24.36 ns/op
BenchmarkGetSingleParallel/ristretto-10   	22281426	        54.99 ns/op
BenchmarkGetSingleParallel/otter-10       	156266952	         7.854 ns/op
BenchmarkSetParallel/theine-10            	 2988525	       398.4 ns/op
BenchmarkSetParallel/ristretto-10         	10083932	       126.0 ns/op
BenchmarkSetParallel/otter-10             	 2239887	       518.2 ns/op
```
Seems I can't reproduce the 
`BenchmarkGetParallel/theine-10         	16778953	        71.68 ns/op`
result(which is much slower than ristretto), is this result from `perf` branch? What is your benchmark(maypok86/benchmarks) result on `perf` branch