@Yiling-J 2024-08-10
@maypok86 I actually prefill the cache in the Get method, but the code hasn't been pushed yet. I've now updated the repo. The Set benchmark also involves some randomness. But I found that I use the total key count as cache size, which will not show the cost of eviction, so I've switched to using 1/10th of the total key count. Here is new result:
```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkGetParallel/theine-10   	38047278	        27.11 ns/op
BenchmarkGetParallel/ristretto-10         	31198701	        38.89 ns/op
BenchmarkGetParallel/otter-10             	113487500	        10.34 ns/op
BenchmarkGetSingleParallel/theine-10      	45519601	        23.73 ns/op
BenchmarkGetSingleParallel/ristretto-10   	22509024	        57.80 ns/op
BenchmarkGetSingleParallel/otter-10       	169000716	         7.001 ns/op
BenchmarkSetParallel/theine-10            	 2272620	       595.4 ns/op
BenchmarkSetParallel/ristretto-10         	10555627	       119.9 ns/op
BenchmarkSetParallel/otter-10             	 2190544	       590.2 ns/op
```
I can't reproduce your Get result(71.68 ns/op), which theine is still much slower than ristretto, is that `perf` branch? What about your benchmark(maypok86/benchmarks) reuslt, this is my result:
```
BenchmarkCache/zipf_otter_reads=100%,writes=0%-8         	94667461	        12.76 ns/op	  78339696 ops/s
BenchmarkCache/zipf_theine_reads=100%,writes=0%-8        	31750724	        32.85 ns/op	  30441332 ops/s
BenchmarkCache/zipf_ristretto_reads=100%,writes=0%-8     	27664843	        44.10 ns/op	  22676619 ops/s
```