@maypok86 2024-08-10
> is that perf branch?

Yeah, that was the first thing I checked. This also proves the change in results since the latest release. For example, here are the results of my benchmarks.

With theine from the perf branch. (`go get github.com/Yiling-J/theine-go@perf`)
```
goos: darwin
goarch: arm64
pkg: github.com/maypok86/benchmarks/throughput
BenchmarkCache/zipf_otter_reads=100%,writes=0%-8                217757455                5.719 ns/op     174842022 ops/s
BenchmarkCache/zipf_theine_reads=100%,writes=0%-8               17539639                64.19 ns/op       15579447 ops/s
BenchmarkCache/zipf_ristretto_reads=100%,writes=0%-8            36242826                32.87 ns/op       30425438 ops/s
PASS
ok      github.com/maypok86/benchmarks/throughput       4.769s
```
With theine from the latest release. (`go get github.com/Yiling-J/theine-go@latest`)
```
goos: darwin
goarch: arm64
pkg: github.com/maypok86/benchmarks/throughput
BenchmarkCache/zipf_otter_reads=100%,writes=0%-8                183928129                6.199 ns/op     161314899 ops/s
BenchmarkCache/zipf_theine_reads=100%,writes=0%-8               10660195               102.0 ns/op         9803268 ops/s
BenchmarkCache/zipf_ristretto_reads=100%,writes=0%-8            40094166                30.01 ns/op       33326931 ops/s
PASS
ok      github.com/maypok86/benchmarks/throughput       4.644s
```