@Yiling-J 2024-08-11
@maypok86 I figured out the issue. The atomic add operation is significantly slower on ARM64 compared to AMD64 (Intel) under high contention. I’ve switched to using the striped counter from xsync instead, which should improve performance on ARM64 and I’ve already tested it on Aliyun Yitian ECS.