@maypok86 2024-08-11
Yes, theine has become much faster.

```
BenchmarkCache/zipf_otter_reads=100%,writes=0%-8                219222543                5.654 ns/op     176863089 ops/s
BenchmarkCache/zipf_theine_reads=100%,writes=0%-8               32655468                38.71 ns/op       25832774 ops/s
BenchmarkCache/zipf_ristretto_reads=100%,writes=0%-8            43702114                31.86 ns/op       31389724 ops/s
```